requirements.txt
pandas
scikit-learn
nltk
streamlit
joblib
PyPDF2
python-docx
fpdf
wordcloud
matplotlib
